Added applyPBC(coords, box) function to core.distances.  This moves all  atoms within the primary unit cell as defined by box.  Will work with both triclinic and orthogonal boxes.
calc_bonds in core.distances now allows triclinic boxes to be used
when calculating minimum image convention.

AtomGroup.packintobox() renamed to packIntoBox for consistency

packIntoBox() now has inplace flag which works similarly to
coordinates.core.convert_to_native().  True changes the coordinates
whereas False just returns them

AtomGroup.bond() now has minimage flag.  If true will account for minimum
image when calculating bond distance.

Lots of AtomGroup methods now have a pbc flag.  If true, this will move
atoms within the primary unit cell before the rest of the calculation.
This is false by default, so default behaviour has not changed.
Affected methods:
centerOfGeometry
centerOfMass
radiusOfGyration
shapeParameter
aspehericity
momentOfInertia
bbox
bsphere
principalAxes

+++ testsuite/MDAnalysisTests/test_atomgroup.py
@@ -216,7 +216,7 @@ class TestAtomGroup(TestCase):
             ag.packintobox()
         assert_raises(ValueError, badpack) #This system has no dimensions, so by default can't pack
 
-        ag.packintobox(box=numpy.array([5.,5.,5.])) #Provide arbitrary box
+        ag.packintobox(box=numpy.array([5.,5.,5.],  dtype=numpy.float32)) #Provide arbitrary box
         assert_array_almost_equal(ag.coordinates()[1000:2000:200],
                                   array([[ 3.94543672,  2.5939188 ,  2.73179913],
                                          [ 3.21632767,  0.879035  ,  0.32085133],
